# docs/backend-contract-pack/openapi.yaml
openapi: 3.1.0
info:
  title: AustriaPath API
  version: 2.0.0-gate0
  description: |
    **Gate 0 — Architecture Freeze** master API contract.
    Aligned with Exam Platform Phases A–G (`1.0.0-phase-g`).

    - Contract pack: `docs/backend-contract-pack/`
    - JSON Schemas: `docs/backend-contract-pack/schemas/`
    - Response envelope: all success responses wrap `data` + `meta`

    **Do not implement divergent endpoints** without contract pack version bump.
servers:
  - url: https://api.austriapath.at/v1
    description: Production
  - url: http://localhost:3000/v1
    description: Local development
...
security:
  - cookieAuth: []
  - bearerAuth: []
paths:
...
  /auth/register:
    post:
      tags: [Auth]
      summary: Register student
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponseEnvelope"
        "409":
          $ref: "#/components/responses/Error"
  /auth/login:
    post:
      tags: [Auth]
      summary: Login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponseEnvelope"
        "401":
          $ref: "#/components/responses/Error"
...
  /auth/forgot-password:
    post:
      tags: [Auth]
      security: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email: { type: string, format: email }
      responses:
        "200":
          description: Always OK
  /auth/reset-password:
    post:
      tags: [Auth]
      security: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [token, password]
              properties:
                token: { type: string }
                password: { type: string, minLength: 8 }
      responses:
        "200":
          description: Password reset
...
  /reports:
    get:
      tags: [Reports]
      summary: List own reports
      parameters:
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 20, maximum: 50 }
        - name: productType
          in: query
          schema:
            $ref: "#/components/schemas/ProductType"
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReportListEnvelope"
...
  /reports/{reportId}:
    get:
      tags: [Reports]
      parameters:
        - name: reportId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FinalReportEnvelope"
...
  /admin/examiner-lab/queue/{labItemId}/resolve:
    post:
      tags: [AdminExaminerLab]
      summary: Resolve lab item
      parameters:
        - name: labItemId
          in: path
          required: true
          schema: { type: string }
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LabResolveRequest"
      responses:
        "200":
          description: Resolved
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: austria_path_session
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      schema: { type: string, format: uuid }
  responses:
    Error:
      description: Error response
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    ProductType:
      type: string
      enum: [placement_test, weekly_plan, ai_exam, intensive_week, premium_month]
...
    ResponseMeta:
      type: object
      required: [requestId, apiVersion, timestamp]
      properties:
        requestId: { type: string, format: uuid }
        apiVersion: { type: string }
        timestamp: { type: string, format: date-time }
    ErrorResponse:
      type: object
      required: [success, error, meta]
      properties:
        success: { const: false }
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string }
            message: { type: string }
            details: { type: object }
        meta:
          $ref: "#/components/schemas/ResponseMeta"
    RegisterRequest:
      type: object
      required: [name, email, password, level]
      properties:
        name: { type: string }
        email: { type: string, format: email }
        password: { type: string, minLength: 8 }
        level: { type: string, enum: [A2, B1, B2] }
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string }
    User:
      type: object
      description: See schemas/user.json
    AuthResponseEnvelope:
      type: object
      properties:
        success: { const: true }
        data:
          type: object
          properties:
            user:
              $ref: "#/components/schemas/User"
        meta:
          $ref: "#/components/schemas/ResponseMeta"
...
    FinalReport:
      type: object
      description: See schemas/final-report.json
...
    FinalReportEnvelope:
      type: object
      properties:
        success: { const: true }
        data:
          type: object
          properties:
            report:
              $ref: "#/components/schemas/FinalReport"
            legacy: { type: object, description: "legacyReportAdapter shape for UI" }
        meta:
          $ref: "#/components/schemas/ResponseMeta"
    ReportListEnvelope:
      type: object
      properties:
        success: { const: true }
        data:
          type: object
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/FinalReport"
            pagination: { type: object }
        meta:
          $ref: "#/components/schemas/ResponseMeta"
...
    LabResolveRequest:
      type: object
      required: [action]
      properties:
        action:
          type: string
          enum: [approve, reject, correct, propose_rule]
        rationale: { type: string }
        correctedDecision: { type: object }
        ruleProposal: { type: object }
...